+++ src/main/java/net/silentchaos512/gear/block/compounder/CompounderBlock.java
@@ -77,8 +77,8 @@ public class CompounderBlock<R extends CompoundingRecipe> extends ModContainerBl
 
     @Override
     public void appendHoverText(ItemStack stack, @Nullable BlockGetter worldIn, List<Component> tooltip, TooltipFlag flagIn) {
-        Set<String> catNameSet = this.info.getCategories().stream().map(IMaterialCategory::getName).collect(Collectors.toSet());
-        String catStr = String.join(", ", catNameSet);
+        Set<Component> catNameSet = this.info.getCategories().stream().map(IMaterialCategory::getDisplayName).collect(Collectors.toSet());
+        Component catStr = TextUtil.separatedList(catNameSet);
         tooltip.add(TextUtil.translate("block", "compounder.desc", catStr));
     }